refactor(tests): streamline imports and improve documentation in test files
