+++ tests/integration/test_simulation_runner.py
@@ -9,9 +9,9 @@ import json
 import logging
 from pathlib import Path
 from typing import Any, Dict, List
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
-import pytest
+from tests.common import Mock, pytest
 from src.data_models.simulation_strategy_config import StrategyConfig
 from src.simulation_runner import SimulationRunner
