+++ tests/type_definitions/flower_typing_reference.py
@@ -18,8 +18,8 @@ from dataclasses import dataclass
 from enum import Enum
 from typing import Any, Callable, Optional, Union
 
-import numpy as np
 import numpy.typing as npt
+from tests.common import np
 
 NDArray = npt.NDArray[Any]
 NDArrayInt = npt.NDArray[np.int_]
