+++ tests/unit/test_data_models/test_client_info.py
@@ -1,4 +1,4 @@
-import pytest
+from tests.common import pytest
 from src.data_models.client_info import ClientInfo
