+++ tests/unit/test_data_models/test_round_info.py
@@ -1,4 +1,4 @@
-import pytest
+from tests.common import pytest
 from src.data_models.round_info import RoundsInfo
 from src.data_models.simulation_strategy_config import StrategyConfig
