+++ tests/unit/test_output_handlers/test_directory_handler.py
@@ -1,8 +1,8 @@
 import csv
 import json
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
-import pytest
+from tests.common import Mock, pytest
 from src.data_models.client_info import ClientInfo
 from src.data_models.round_info import RoundsInfo
 from src.data_models.simulation_strategy_config import StrategyConfig
