+++ tests/unit/test_simulation_strategies/test_krum_strategy.py
@@ -4,18 +4,15 @@ Unit tests for KrumBasedRemovalStrategy.
 Tests Krum client selection algorithms, distance calculations, and subset identification.
 """
 
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
-import numpy as np
-import pytest
-from flwr.common import FitRes, ndarrays_to_parameters
-from flwr.server.client_proxy import ClientProxy
+from tests.common import Mock, np, pytest, FitRes, ndarrays_to_parameters, ClientProxy
 from src.data_models.simulation_strategy_history import SimulationStrategyHistory
 from src.simulation_strategies.krum_based_removal_strategy import (
     KrumBasedRemovalStrategy,
 )
 
-from tests.conftest import generate_mock_client_data
+from tests.common import generate_mock_client_data
 
 
 class TestKrumBasedRemovalStrategy:
