+++ tests/unit/test_simulation_strategies/test_rfa_strategy.py
@@ -4,15 +4,20 @@ Unit tests for RFABasedRemovalStrategy.
 Tests RFA (Robust Federated Averaging) geometric median calculation and client removal logic.
 """
 
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
-import numpy as np
-import pytest
-from flwr.common import FitRes, ndarrays_to_parameters, parameters_to_ndarrays
-from flwr.server.client_proxy import ClientProxy
+from tests.common import (
+    Mock,
+    np,
+    pytest,
+    FitRes,
+    ndarrays_to_parameters,
+    parameters_to_ndarrays,
+    ClientProxy,
+)
 from src.simulation_strategies.rfa_based_removal_strategy import RFABasedRemovalStrategy
 
-from tests.conftest import generate_mock_client_data
+from tests.common import generate_mock_client_data
 
 
 class TestRFABasedRemovalStrategy:
